move state mgmt, handlers into parent & add props to child
