+++ client/src/components/ProjectWizard/WizardFooter.jsx
@@ -1,5 +1,4 @@
-import React, { useState, useRef, useContext } from "react";
-import { useNavigate } from "react-router-dom";
+import React, { useRef, useContext } from "react";
 import PropTypes from "prop-types";
 import NavButton from "../Button/NavButton";
 import { createUseStyles } from "react-jss";
@@ -10,20 +9,6 @@ import { formatDatetime } from "../../helpers/util";
 import UserContext from "../../contexts/UserContext";
 import Popup from "reactjs-popup";
 import ProjectContextMenu from "../Projects/ProjectContextMenu";
-import useErrorHandler from "../../hooks/useErrorHandler";
-
-import CsvModal from "../Modals/ActionProjectsCsv";
-import WarningProjectDelete from "../Modals/WarningProjectDelete";
-import SnapshotProjectModal from "../Modals/ActionProjectSnapshot";
-import RenameSnapshotModal from "../Modals/ActionSnapshotRename";
-import ShareSnapshotModal from "../Modals/ActionSnapshotShare";
-import WarningSnapshotSubmit from "../Modals/WarningSnapshotSubmit";
-import InfoTargetNotReached from "../Modals/InfoTargetNotReached";
-import InfoSnapshotSubmit from "components/Modals/InfoSnapshotSubmitted";
-import CopyProjectModal from "../Modals/ActionProjectCopy";
-
-import * as projectService from "../../services/project.service";
-import * as projectResultService from "../../services/projectResult.service";
 
 const useStyles = createUseStyles({
   allButtonsWrapper: {
@@ -69,6 +54,14 @@ const WizardFooter = ({
   onSave,
   showCopyAndEditSnapshot,
   showSubmitModal,
+  handleCsvModalOpen,
+  handleHide,
+  handleDeleteModalOpen,
+  handlePrintPdf,
+  handleSnapshotModalOpen,
+  handleRenameSnapshotModalOpen,
+  handleShareSnapshotModalOpen,
+  handleSubmitModalOpen,
   project,
   shareView
 }) => {
@@ -84,21 +77,6 @@ const WizardFooter = ({
   const userContext = useContext(UserContext);
   const loggedInUserId = userContext.account?.id || null;
 
-  const email = userContext.account ? userContext.account.email : "";
-  const navigate = useNavigate();
-  const handleError = useErrorHandler(email, navigate);
-
-  const [snapshotModalOpen, setSnapshotModalOpen] = useState(false);
-  const [submitModalOpen, setSubmitModalOpen] = useState(false);
-  const [successModelOpen, setSuccessModelOpen] = useState(false);
-  const [targetNotReachedModalOpen, setTargetNotReachedModalOpen] =
-    useState(false);
-  const [renameSnapshotModalOpen, setRenameSnapshotModalOpen] = useState(false);
-  const [deleteModalOpen, setDeleteModalOpen] = useState(false);
-  const [csvModalOpen, setCsvModalOpen] = useState(false);
-  const [shareSnapshotModalOpen, setShareSnapshotModalOpen] = useState(false);
-
-  const [copyModalOpen, setCopyModalOpen] = useState(false);
   const isAdmin = userContext.account?.isAdmin || false;
 
   const setDisabledSubmitButton = () => {
@@ -126,188 +104,8 @@ const WizardFooter = ({
     return false;
   };
 
-  const handleRenameSnapshotModalOpen = () => {
-    setRenameSnapshotModalOpen(true);
-  };
-
-  const handleRenameSnapshotModalClose = async (action, newProjectName) => {
-    if (action === "ok") {
-      try {
-        await projectService.renameSnapshot({
-          id: project.id,
-          name: newProjectName
-        });
-      } catch (err) {
-        handleError(err);
-      }
-    }
-    setRenameSnapshotModalOpen(false);
-  };
-
-  const handleShareSnapshotModalOpen = () => {
-    setShareSnapshotModalOpen(true);
-  };
-
-  const handleShareSnapshotModalClose = async () => {
-    setShareSnapshotModalOpen(false);
-  };
-
-  const handleSubmitModalOpen = () => {
-    if (project.earnedPoints >= project.targetPoints) {
-      setSubmitModalOpen(true);
-    } else {
-      setTargetNotReachedModalOpen(true);
-    }
-  };
-
-  const handleSubmitModalClose = async action => {
-    if (action === "ok") {
-      // await updateProjects();
-      setSuccessModelOpen(true);
-    }
-    setSubmitModalOpen(false);
-  };
-
-  const handleCsvModalOpen = () => {
-    setCsvModalOpen(true);
-  };
-
-  const handleCsvModalClose = async () => {
-    setCsvModalOpen(false);
-  };
-
-  const handleSnapshotModalClose = async () => {
-    // if (action === "ok") {
-    //   try {
-    //     await projectService.snapshot({
-    //       id: selectedProject.id,
-    //       name: newProjectName
-    //     });
-    //     await updateProjects();
-    //     setSelectedProject(null);
-    //   } catch (err) {
-    //     handleError(err);
-    //   }
-    // }
-    setSnapshotModalOpen(false);
-  };
-
-  const handleSnapshotModalOpen = () => {
-    setSnapshotModalOpen(true);
-  };
-
-  const handleSuccessModalClose = () => {
-    setSuccessModelOpen(false);
-  };
-
-  const handleCopyModalOpen = () => {
-    setCopyModalOpen(true);
-  };
-
-  const handleCopyModalClose = async (action, newProjectName) => {
-    let newSelectedProject = { ...project };
-    if (action === "ok") {
-      const projectFormInputsAsJson = JSON.parse(project.formInputs);
-      projectFormInputsAsJson.PROJECT_NAME = newProjectName;
-      if (!project.targetPoints) {
-        await projectResultService.populateTargetPoints(project);
-        newSelectedProject = await projectService.getById(project.id);
-      }
-      let newProject = {
-        ...newSelectedProject,
-        loginId: loggedInUserId,
-        name: newProjectName,
-        formInputs: JSON.stringify(projectFormInputsAsJson)
-      };
-      if (!newProject.description) {
-        newProject.description = "";
-      }
-      try {
-        await projectService.post(newProject);
-      } catch (err) {
-        handleError(err);
-      }
-    }
-    setCopyModalOpen(false);
-  };
-
-  const handleHide = async project => {
-    try {
-      const projectIDs = [project.id];
-      const dateHidden = !project.dateHidden;
-
-      await projectService.hide(projectIDs, dateHidden);
-    } catch (err) {
-      console.error(err);
-    }
-  };
-
-  const handleDeleteModalOpen = () => {
-    setDeleteModalOpen(true);
-  };
-
-  const handleDeleteModalClose = async action => {
-    if (action === "ok") {
-      const projectIDs = [project.id];
-      const dateTrashed = !project.dateTrashed;
-      try {
-        await projectService.trash(projectIDs, dateTrashed);
-      } catch (err) {
-        handleError(err);
-      }
-    }
-    setDeleteModalOpen(false);
-  };
-
   return (
     <>
-      <>
-        <CsvModal
-          mounted={csvModalOpen}
-          onClose={handleCsvModalClose}
-          project={project}
-        />
-        <CopyProjectModal
-          mounted={copyModalOpen}
-          onClose={handleCopyModalClose}
-          selectedProjectName={project.name}
-        />
-        <WarningProjectDelete
-          mounted={deleteModalOpen}
-          onClose={handleDeleteModalClose}
-          project={project}
-        />
-        <SnapshotProjectModal
-          mounted={snapshotModalOpen}
-          onClose={handleSnapshotModalClose}
-          selectedProjectName={project}
-        />
-        <RenameSnapshotModal
-          mounted={renameSnapshotModalOpen}
-          onClose={handleRenameSnapshotModalClose}
-          selectedProjectName={project.name}
-        />
-        <ShareSnapshotModal
-          mounted={shareSnapshotModalOpen}
-          onClose={handleShareSnapshotModalClose}
-          project={project}
-        />
-        <WarningSnapshotSubmit
-          mounted={submitModalOpen}
-          onClose={handleSubmitModalClose}
-          project={project}
-        />
-        <InfoTargetNotReached
-          mounted={targetNotReachedModalOpen}
-          onClose={() => setTargetNotReachedModalOpen(false)}
-          project={project}
-        />
-        <InfoSnapshotSubmit
-          mounted={successModelOpen}
-          onClose={handleSuccessModalClose}
-          project={project}
-        />
-      </>
       <div id="all-buttons-wrapper" className={classes.allButtonsWrapper}>
         {rules && rules.length ? ( //navigation disabled until rules have loaded
           <>
@@ -367,12 +165,11 @@ const WizardFooter = ({
                     <ProjectContextMenu
                       project={project}
                       closeMenu={close}
-                      handleCsvModalOpen={ev => handleCsvModalOpen(ev, project)}
+                      handleCsvModalOpen={handleCsvModalOpen}
                       handleHide={handleHide}
                       handleDeleteModalOpen={handleDeleteModalOpen}
-                      handlePrintPdf={null}
+                      handlePrintPdf={handlePrintPdf}
                       handleSnapshotModalOpen={handleSnapshotModalOpen}
-                      handleCopyModalOpen={handleCopyModalOpen}
                       handleRenameSnapshotModalOpen={
                         handleRenameSnapshotModalOpen
                       }
@@ -485,6 +282,13 @@ WizardFooter.propTypes = {
   targetNotReachedModalOpen: PropTypes.any,
   showCopyAndEditSnapshot: PropTypes.func,
   showSubmitModal: PropTypes.func,
+  handleCsvModalOpen: PropTypes.func,
+  handleHide: PropTypes.func,
+  handleDeleteModalOpen: PropTypes.func,
+  handlePrintPdf: PropTypes.func,
+  handleSnapshotModalOpen: PropTypes.func,
+  handleRenameSnapshotModalOpen: PropTypes.func,
+  handleShareSnapshotModalOpen: PropTypes.func,
   onDownload: PropTypes.any,
   project: PropTypes.any,
   selectProject: PropTypes.any,
